chore: move lucide icons to asset folder uninstall lucide-react
