+++ package.json
@@ -18,7 +18,6 @@
     "clsx": "^2.0.0",
     "date-fns": "^2.30.0",
     "localforage": "^1.10.0",
-    "lucide-react": "^0.260.0",
     "match-sorter": "^6.3.1",
     "react": "^18.2.0",
     "react-day-picker": "^8.8.0",
