+++ src/components/CustomDropdown.tsx
@@ -1,5 +1,5 @@
-import Icon from '@components/ui/Icon';
 import { ReactNode, forwardRef, useEffect, useState } from 'react';
+import iconCheck from '@assets/icon-check.svg'
 
 // type DropdownRefProps = {
 //     ref: React.RefObject<HTMLDivElement>,
@@ -82,9 +82,17 @@ const CustomDropdown = forwardRef<HTMLDivElement, CustomDropdownProps>(
 										className={`flex items-center gap-x-2.5 py-4 px-2 text-sm rounded-sm outline-none outline-0 ring-0 hover:bg-slate-100 focus:bg-slate-100 w-full`}
 										onClick={() => selectDropdown(optionValue)}
 									>
-										<Icon name='check' className={`h-4 w-4 ${
-												options.selected === optionValue ? '' : 'opacity-0'
-											}`} />
+										<div className="h-4 w-4">
+											<img
+												src={iconCheck}
+												alt=""
+												className={
+													options.selected === optionValue
+														? ''
+														: 'hidden'
+												}
+											/>
+										</div>
 										<p className="text-left">{optionValue}</p>
 									</button>
 								</li>
