+++ src/components/order-detail/AddFeeModal.tsx
@@ -2,8 +2,8 @@
 import { useEffect, useRef, useState } from 'react';
 import useTrackedModalStore from '@stores/modalStore';
 import { FeeType, useTrackedOrderStore } from '@stores/orderStore.tsx';
-import Icon from '@components/ui/Icon';
 import iconClose from '@assets/icon-x.svg';
+import iconArrowDown from '@assets/icon-arrow-down.svg'
 import iconExclamation from '@assets/icon-exclamation-circle.svg';
 import CustomDropdown from '../CustomDropdown.tsx';
 
@@ -144,8 +144,9 @@ const AddFeeModal = () => {
 					}
 				>
 					<p>{category}</p>
-					<Icon
-						name="chevron-down"
+					<img
+						src={iconArrowDown}
+						alt=""
 						className="h-4 w-4 opacity-50 absolute right-3 bottom-1/2 translate-y-1/2"
 					/>
 				</button>
